Fix issues found by static analysis (#183)
* Fix bad integer size for indexes, which are able to cause overflow

* Move ini implementation to single file, for easy exclusion in semmle

* Don't open files without specifying a umask

* Fix some issues with NULL deref errors

* Fix NULL check for version

* Remove useless GetAutoFailoverNode

* Make sure pid does not overflow

* Make sure pgport does not overflow

* Handle malloc returning NULL

+++ src/bin/pg_autoctl/pgsetup.c
@@ -437,7 +437,7 @@ get_pgpid(PostgresSetup *pgSetup, bool pg_is_not_running_is_ok)
 
  	fclose(fp);
 
-	if (pid > 0)
+	if (pid > 0 && pid <= INT_MAX)
 	{
 		if (kill(pid, 0) == 0)
 		{
@@ -1111,13 +1111,14 @@ int
 pgsetup_get_pgport()
 {
 	char *pgport_env = getenv("PGPORT");
-	int pgport = 0;
+	long pgport = 0;
 
 	if (pgport_env)
 	{
+		errno = 0;
 		pgport = strtol(pgport_env, NULL, 10);
 
-		if (pgport > 0 && errno != EINVAL)
+		if (pgport > 0 && pgport < INT_MAX && errno == 0)
 		{
 			return pgport;
 		}
